Reorg to put protofix executable under /cmd
