+++ .gitignore
@@ -4,7 +4,7 @@
 
 # Build artifacts
 tmp-test-pod.yml
-protofix
+/protofix
 
 # Compiled Object files, Static and Dynamic libs (Shared Objects)
 *.o
