+++ src/pkg/protofix/example1_output_pb_go
