+++ src/pkg/protofix/example1_pb_go
